Added two new fields for custom mods, and errorproofing
In anticipation of changes to the Emergency template as well as the BNDX
scrubber, I've added two fields for modifying those scripts.

There's also some error handling, including checks for the above to
fields to make sure the user doesn't add stray characters, as well as
the custom file path (it'll now warn you if you put a slash at the end
of your file path).