Resolved #2
It now contains a new tab for zip/download bypass options, and will
simply copy the directory in agencies where they do not have the ability
to use the program's zip/download features (some of that is .NET
incompatibility, some is network connectivity issues, some is issues
with the unzip.vbs workaround for .NET 4.0 and before). This should
allow lots of folks to use most features of the installer, without
impacting the majority of whom there are no issues.